Se modificó el menú de uart
Se modificó el menu de Uart, debido a que preferí simplificar, hace un par de opciones más de navegación y simplificarme la vida. Ahora ya solo pide elegir el servomotor y posteriormente agregar el ángulo de 0 a 180. Esto lo hice debido a que no quería hacer más programación defensiva.

+++ P2_Animatronic_/P2_Animatronic_/PWM2F/PWM2.h
@@ -12,7 +12,8 @@
 #include <stdint.h>
 
 void PWM2_Init(void);
-void PWM2_SetDuty1(uint8_t ticks);
-void PWM2_SetDuty2(uint8_t ticks);
+
+void PWM2_SetDuty1(uint16_t ticks);
+void PWM2_SetDuty2(uint16_t ticks);
 
 #endif /* PWM2_H_ */
